requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.6.12
matplotlib>=3.5.0
numpy>=1.20.0
Pillow>=9.0.0
scikit-learn>=1.0.0
tqdm>=4.60.0
requests>=2.25.0
wandb>=0.15.0
seaborn>=0.12.0 